refactor(direct-control): extract progression closeout types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only technology closeout input/result types into packages/civ7-direct-control/src/play/progression/technology.ts and culture closeout input/result types into packages/civ7-direct-control/src/play/progression/culture.ts. Keeps public facade type re-exports in index.ts and leaves the public request wrapper ownership in index.ts.

Preserves public type contracts and does not change technology/culture closeout source strings, command serialization, parser labels, wrapper ownership, mutation behavior, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Diplomacy/narrative closeout, operation, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- technology-choice-closeout.test.ts culture-choice-closeout.test.ts progression-reads.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.technology.test.ts game.play.culture.test.ts game.play.progression-read.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- progression closeout type ownership scan and progression module no-cycle import scan

Local package/type relocation proof only; no runtime/live-game, closeout behavior, CLI semantic projection, telemetry, AI-ingestion, hotseat product-path, or procedure-core proof claimed.
