+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -766,6 +766,19 @@ All future agent waves must be framed before delegation:
   telemetry type ownership remain pending separate owner slices. This is local
   package/type relocation proof only, not runtime proof, AI/hotseat product-path
   support, or procedure-core readiness.
+- Direct-control progression chooser closeout types slice: completed as a
+  narrow public type ownership relocation. It moves only technology closeout
+  input/result types into `src/play/progression/technology.ts` and culture
+  closeout input/result types into `src/play/progression/culture.ts` while
+  keeping facade type re-exports in `index.ts`. This preserves public type
+  contracts and does not change technology/culture closeout source strings,
+  command serialization, parser labels, wrapper ownership, mutation behavior,
+  runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI
+  semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4
+  matrix-row acceptance. Diplomacy/narrative closeout, operation, public
+  procedure schema, and telemetry type ownership remain pending separate owner
+  slices. This is local package/type relocation proof only, not runtime proof,
+  AI/hotseat product-path support, or procedure-core readiness.
 - Direct-control turn-completion slice: completed as a narrow turn-completion
   source/wrapper relocation. It moves only `getCiv7TurnCompletionStatus`,
   `sendCiv7TurnComplete`, `sendCiv7TurnUnready`, the status command builder,
